** Moved timed notification strategies into their own package
